guards for not logged users and not creators

+++ Javascript Back End/Cubicle/controllers/deleteController.js
@@ -1,3 +1,4 @@
+const { getAttachedAccessories } = require('../services/accessoryService.js');
 const { getCubeById, deleteCube } = require('../services/cubeService.js');
 const { getSelectedOption } = require('./utils/utils.js');
 
@@ -7,6 +8,16 @@ deleteController.get('/:id', async (req, res) => {
     const id = req.params.id;
     const cube = await getCubeById(id);
     const options = getSelectedOption(cube.difficultyLevel);
+    const attachedAccessories = await getAttachedAccessories(cube);
+
+    if (!req.user || cube.ownerId != req.user._id) {
+        return res.render('details', {
+            cube,
+            attachedAccessories,
+            title: 'Details',
+            error: "You're not creator of this cube!",
+        });
+    }
 
     res.render('delete', {
         title: 'Delete',
@@ -19,6 +30,16 @@ deleteController.post('/:id', async (req, res) => {
     const id = req.params.id;
     const cube = await getCubeById(id);
     const options = getSelectedOption(cube.difficultyLevel);
+    const attachedAccessories = await getAttachedAccessories(cube);
+
+    if (!req.user || cube.ownerId != req.user._id) {
+        return res.render('details', {
+            cube,
+            attachedAccessories,
+            title: 'Details',
+            error: "You're not creator of this cube!",
+        });
+    }
 
     try {
         await deleteCube(id);
